Move galaxy.util.log to galaxy.util.logging
