+++ lib/galaxy/app.py
@@ -226,7 +226,7 @@ class UniverseApplication(object, config.ConfiguresGalaxyMixin):
 
     def configure_fluent_log(self):
         if self.config.fluent_log:
-            from galaxy.util.log.fluent_log import FluentTraceLogger
+            from galaxy.util.logging.fluent_log import FluentTraceLogger
             self.trace_logger = FluentTraceLogger('galaxy', self.config.fluent_host, self.config.fluent_port)
         else:
             self.trace_logger = None
